# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentrify API",
        "description": "API documentation for the Mentrify platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Mentrify API Server"
        }
    ],
    "paths": {
...
        "/api/mentee/dashboard": {
            "get": {
                "tags": [
                    "Mentee Dashboard"
                ],
                "summary": "Get mentee dashboard data",
                "description": "Retrieves all necessary data for the mentee dashboard",
                "operationId": "getMenteeDashboard",
                "responses": {
                    "200": {
                        "description": "Mentee dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "active_mentors_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "learning_hours": {
                                            "type": "integer",
                                            "example": 16
                                        },
                                        "goals_completed_count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "upcoming_sessions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Session"
                                            }
                                        },
                                        "past_sessions": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Session"
                                                    },
                                                    {
                                                        "properties": {
                                                            "feedback_given": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "first_name": {
                        "description": "First name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string",
                        "format": "email"
                    },
                    "user_type": {
                        "description": "Type of user",
                        "type": "string",
                        "enum": [
                            "mentor",
                            "mentee"
                        ]
                    },
                    "photo_url": {
                        "description": "URL of the user's photo",
                        "type": "string",
                        "nullable": true
                    },
                    "job_title": {
                        "description": "Job title (for mentors)",
                        "type": "string",
                        "nullable": true
                    },
                    "category": {
                        "description": "Category (for mentors)",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Session": {
                "title": "Session",
                "description": "Mentoring session model",
                "properties": {
                    "id": {
                        "description": "Session ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "mentor_id": {
                        "description": "ID of the Mentor (User)",
                        "type": "integer",
                        "format": "int64"
                    },
                    "mentee_id": {
                        "description": "ID of the Mentee (User)",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Title or main topic of the session",
                        "type": "string"
                    },
                    "description": {
                        "description": "Detailed description or agenda",
                        "type": "string",
                        "nullable": true
                    },
                    "start_time": {
                        "description": "Scheduled start time of the session",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "description": "Scheduled end time of the session",
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "description": "Status of the session",
                        "type": "string",
                        "enum": [
                            "upcoming",
                            "completed",
                            "cancelled",
                            "in_progress"
                        ]
                    },
                    "join_url": {
                        "description": "URL to join the virtual session",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp of session creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of last session update",
                        "type": "string",
                        "format": "date-time"
                    },
                    "mentor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "mentee": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}